[spac project] missile modifiers

+++ PHEngine/PHEngine/src/Core/GameCore/Event/PhysicsCollisionOccuredEvent.h
@@ -12,9 +12,9 @@ namespace Event
 {
    using ParentEvent_t = TEvent<SingleDataEventPolicy<ePhysicsBodyType,
                                                       uint32_t /*this_phys_descriptor_id*/,
-                                                      uint32_t /*this_actor_id*/,
+                                                      uint64_t /*this_actor_id*/,
                                                       uint32_t /*that_phys_descriptor_id*/,
-                                                      uint32_t /*that_actor_id*/>>;
+                                                      uint64_t /*that_actor_id*/>>;
 
    class PhysicsCollisionOccuredEvent
        : public ParentEvent_t
